Move `ReadReplicaConnectionCreationException` to the API
