+++ CHANGELOG.md
@@ -24,6 +24,9 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://github.com/atlassian-labs/db-replica/compare/release-2.8.0...master
 
+### Added
+- `ReadReplicaConnectionCreationException`
+
 ## [2.8.0] - 2022-10-04
 [2.8.0]: https://github.com/atlassian-labs/db-replica/compare/release-2.7.2...release-2.8.0
